feat(payment): refactor payment methods handling and update structure
- Consolidate payment method icons and labels into a single source.
- Update PaymentDialog and PaymentStatusDialog components to utilize new payment methods structure.
- Remove deprecated data file for payment methods.
- Extend PaymentMethod type to include 'debit'.
