+++ src/features/transactions/components/transactions-columns.tsx
@@ -2,7 +2,7 @@ import { ColumnDef } from '@tanstack/react-table'
 import { cn } from '@/lib/utils'
 import { Checkbox } from '@/components/ui/checkbox'
 import { formatCurrency } from '@/utils/format'
-import { paymentMethods } from '../data/data'
+import { paymentMethods } from '@/lib/payment-methods'
 import { Transaction, TransactionPaymentMethod } from '@/types/transactions'
 import { DataTableColumnHeader } from './data-table-column-header'
 import { DataTableRowActions } from './data-table-row-actions'
